Fix: map component was rendering hence zooming out constantly
